src/dynamic_infrastructure.py
import pulumi
from pulumi_aws import ec2, autoscaling, lb, cloudformation

def create_infrastructure():
    # Reference the CloudFormation stack
    vpc_stack = cloudformation.get_stack(name="ProxyVPCStack")

    # Get outputs from the CloudFormation stack
    vpc_id = vpc_stack.outputs["VpcId"]
    public_subnet_1_id = vpc_stack.outputs["PublicSubnet1"]
    public_subnet_2_id = vpc_stack.outputs["PublicSubnet2"]

    # Configuration
    config = pulumi.Config()
    instance_type = config.get("instanceType") or "t3.micro"
    key_name = config.require("keyName")
    ami_id = config.require("amiId")  # Make sure to provide this in your Pulumi config

    # Security Group
    security_group = ec2.SecurityGroup("proxy-security-group",
        description="Security group for proxy servers",
        vpc_id=vpc_id,
        ingress=[
            ec2.SecurityGroupIngressArgs(protocol="tcp", from_port=22, to_port=22, cidr_blocks=["0.0.0.0/0"]),
            ec2.SecurityGroupIngressArgs(protocol="tcp", from_port=3128, to_port=3128, cidr_blocks=["0.0.0.0/0"])
        ])

    # Launch Template
    user_data = """#!/bin/bash
yum update -y
yum install -y squid
systemctl enable squid
systemctl start squid
"""

    launch_template = ec2.LaunchTemplate("proxy-launch-template",
        image_id=ami_id,
        instance_type=instance_type,
        key_name=key_name,
        vpc_security_group_ids=[security_group.id],
        user_data=pulumi.Output.from_input(user_data).apply(lambda ud: ud.encode().decode('ascii')))

    # Auto Scaling Group
    asg = autoscaling.Group("proxy-asg",
        vpc_zone_identifiers=[public_subnet_1_id, public_subnet_2_id],
        launch_template=autoscaling.GroupLaunchTemplateArgs(
            id=launch_template.id,
            version="$Latest"
        ),
        min_size=500,
        max_size=1000,
        desired_capacity=500,
        health_check_type="EC2",
        health_check_grace_period=300)

    # Network Load Balancer
    load_balancer = lb.LoadBalancer("proxy-load-balancer",
        load_balancer_type="network",
        subnets=[public_subnet_1_id, public_subnet_2_id])

    target_group = lb.TargetGroup("proxy-target-group",
        port=3128,
        protocol="TCP",
        vpc_id=vpc_id,
        target_type="instance")

    listener = lb.Listener("proxy-listener",
        load_balancer_arn=load_balancer.arn,
        port=3128,
        protocol="TCP",
        default_actions=[lb.ListenerDefaultActionArgs(
            type="forward",
            target_group_arn=target_group.arn
        )])

    # Outputs
    pulumi.export("proxy_load_balancer_dns", load_balancer.dns_name)

# This is the entry point for the Pulumi program
if __name__ == "__main__":
    create_infrastructure()